rename testCommon.js to test-common.js
