+++ abstract/put-get-del-test.js
@@ -168,7 +168,7 @@ module.exports.tearDown = function (test, testCommon) {
 }
 
 module.exports.all = function (leveldown, test, testCommon) {
-  testCommon = testCommon || require('../testCommon')
+  testCommon = testCommon || require('../test-common')
   module.exports.setUp(leveldown, test, testCommon)
   module.exports.errorKeys(test)
   module.exports.nonErrorKeys(test)
