+++ test.js
@@ -3,7 +3,7 @@
 var test = require('tape')
 var sinon = require('sinon')
 var inherits = require('util').inherits
-var testCommon = require('./testCommon')
+var testCommon = require('./test-common')
 var AbstractLevelDOWN = require('./').AbstractLevelDOWN
 var AbstractIterator = require('./').AbstractIterator
 var AbstractChainedBatch = require('./').AbstractChainedBatch
